Major Styling Changes
This commit has virtually renamed every class in the theme. The original theme did not resize how I would have liked on mobile or tablet devices. As a result,  I have changed some things. The stylesheet in this commit is far from complete but the path forward is now more clear. With what is now an easier to understand stylesheet, I hope to make significant progress in adding a responsive design to the theme. I have also removed some of the unecesssary dependencies.